g2g utils moved to gridtile.py module to remove dependance of field and plotting scripts on f2py.
