+++ GEOS_Util/coupled_diagnostics/g5lib/utils.py
@@ -1,9 +1,10 @@
 '''
 Different routines used by field, dset, grid etc.
+
+Note: g2g functionality was moved to gridtile.py due to dependence on f2py.
 '''
 
 import scipy as sp
-import futils
 
 def scalar2slice(*args):
     val=[]
@@ -52,31 +53,3 @@ def gint(y, x, axis=-1):
     slice2[axis] = slice(None,-1)
 
     return sp.ma.sum(d*(y[slice1]+y[slice2])/2.0,axis)
-
-
-def g2g(datain,tiledata,hdimsout):
-    '''
-    Interpolates field from one horizontal grid onto another using exchange 
-    grid (tiles).
-    
-    datain - input array (at least 2d)
-    tiledata - array of records of size Nt (Nt - number of tiles).
-    tiledata['iin'], tiledata['jin'], tiledata['iout'], 
-    tiledata['jout'], tiledata['frac']  - 
-    arrays of size Nt of inexes on the input grid, indexes on the output grid, 
-    corresponding to a tile N, and fraction of tile on the output grid
-    hdimsout - (imout,jmout), dimensions of output grid
-    '''
-    shin=datain.shape
-    shout=datain.shape[:-2]+hdimsout
-            
-    return sp.ma.masked_values(
-        futils.g2g(datain.filled().reshape((-1,)+shin[-2:]),
-                   tiledata['iin'].astype(int),
-                   tiledata['jin'].astype(int),
-                   tiledata['iout'].astype(int),
-                   tiledata['jout'].astype(int),
-                   tiledata['frac'],
-                   datain.fill_value,
-                   shout[-1],
-                   shout[-2]),datain.fill_value).reshape(shout)
